@@ -7,13 +7,13 @@
 
 class AbstractJobsController < ApplicationController
 
-  before_filter :authenticate_user!
+  before_action :authenticate_user!
 
   def index
     @abstract_jobs = AbstractJob.search(params)
-    response.headers["X-total"] = @abstract_jobs.total_count.to_s
-    response.headers["X-offset"] = @abstract_jobs.offset_value.to_s
-    response.headers["X-limit"] = @abstract_jobs.limit_value.to_s
+    response.headers['X-total'] = @abstract_jobs.total_count.to_s
+    response.headers['X-offset'] = @abstract_jobs.offset_value.to_s
+    response.headers['X-limit'] = @abstract_jobs.limit_value.to_s
     render json: @abstract_jobs, serializer: ActiveModel::ArraySerializer
   end
 
