@@ -1,13 +1,13 @@
-# The Supplejack Worker code is Crown copyright (C) 2014, New Zealand Government, 
-# and is licensed under the GNU General Public License, version 3. 
-# See https://github.com/DigitalNZ/supplejack_worker for details. 
-# 
+# The Supplejack Worker code is Crown copyright (C) 2014, New Zealand Government,
+# and is licensed under the GNU General Public License, version 3.
+# See https://github.com/DigitalNZ/supplejack_worker for details.
+#
 # Supplejack was created by DigitalNZ at the National Library of NZ
 # and the Department of Internal Affairs. http://digitalnz.org/supplejack
 
 class HarvestJobsController < ApplicationController
 
-  before_filter :authenticate_user!
+  before_action :authenticate_user!
 
   def index
     @harvest_jobs = HarvestJob.search(params)
@@ -23,18 +23,24 @@ class HarvestJobsController < ApplicationController
   end
 
   def create
-    @harvest_job = HarvestJob.new(params[:harvest_job])
-    
+    @harvest_job = HarvestJob.new(harvest_job_params)
+
     if @harvest_job.save
       render json: @harvest_job
     else
-      render json: {errors: @harvest_job.errors }, status: 422
+      render json: { errors: @harvest_job.errors }, status: 422
     end
   end
 
   def update
     @harvest_job = HarvestJob.find(params[:id])
-    @harvest_job.update_attributes(params[:harvest_job])
+    @harvest_job.update_attributes(harvest_job_params)
     render json: @harvest_job
+  end
+
+  private
+
+  def harvest_job_params
+    params.require(:harvest_job).permit!
   end
 end
