@@ -1,11 +1,11 @@
-# The Supplejack Worker code is Crown copyright (C) 2014, New Zealand Government, 
-# and is licensed under the GNU General Public License, version 3. 
-# See https://github.com/DigitalNZ/supplejack_worker for details. 
-# 
+# The Supplejack Worker code is Crown copyright (C) 2014, New Zealand Government,
+# and is licensed under the GNU General Public License, version 3.
+# See https://github.com/DigitalNZ/supplejack_worker for details.
+#
 # Supplejack was created by DigitalNZ at the National Library of NZ
 # and the Department of Internal Affairs. http://digitalnz.org/supplejack
 
-require "spec_helper"
+require 'rails_helper'
 
 describe HarvestJobsController do
 
@@ -14,46 +14,48 @@ describe HarvestJobsController do
   before(:each) do
     controller.stub(:authenticate_user!) { true }
   end
-  
-  describe "POST create" do
+
+  describe 'POST create' do
     before(:each) do
       HarvestJob.stub(:new) { job }
     end
 
-    it "initializes a new harvest job" do
-      HarvestJob.should_receive(:new).with({"strategy" => "xml", "file_name" => "youtube.rb"}) { job }
-      post :create, harvest_job: {strategy: "xml", file_name: "youtube.rb"}, format: "js"
-      assigns(:harvest_job).should eq job
+    it 'initializes a new harvest job' do
+      params = ActionController::Parameters.new(strategy: 'xml', file_name: 'youtube.rb').permit!
+      HarvestJob.should_receive(:new).with(params) { job }
+      post :create, params: { harvest_job: {strategy: 'xml', file_name: 'youtube.rb'} }, format: 'js'
+      expect(assigns(:harvest_job)).to eq job
     end
 
-    it "should save the harvest job" do
+    it 'should save the harvest job' do
       job.should_receive(:save)
-      post :create, harvest_job: {strategy: "xml", file_name: "youtube.rb"}, format: "js"
+      post :create, params: { harvest_job: {strategy: 'xml', file_name: 'youtube.rb'} }, format: 'js'
     end
   end
 
-  describe "#GET show" do
-    it "finds the harvest job" do
-      HarvestJob.should_receive(:find).with("1") { job }
-      get :show, id: 1, format: "js"
-      assigns(:harvest_job).should eq job
+  describe '#GET show' do
+    it 'finds the harvest job' do
+      HarvestJob.should_receive(:find).with('1') { job }
+      get :show, params: { id: 1 }, format: 'js'
+      expect(assigns(:harvest_job)).to eq job
     end
   end
 
-  describe "PUT Update" do
+  describe 'PUT Update' do
     before(:each) do
-      HarvestJob.stub(:find).with("1") { job }
+      HarvestJob.stub(:find).with('1') { job }
     end
 
-    it "finds the harvest job" do
-      HarvestJob.should_receive(:find).with("1") { job }
-      put :update, id: 1, format: "js"
-      assigns(:harvest_job).should eq job
+    it 'finds the harvest job' do
+      HarvestJob.should_receive(:find).with('1') { job }
+      put :update, params: { id: 1, harvest_job: { stop: true } }, format: 'js'
+      expect(assigns(:harvest_job)).to eq job
     end
 
-    it "should update the attributes" do
-      job.should_receive(:update_attributes).with({"stop" => true})
-      put :update, id: 1, harvest_job: {stop: true}, format: "js"
+    it 'should update the attributes' do
+      params = ActionController::Parameters.new(stop: 'true').permit!
+      job.should_receive(:update_attributes).with(params)
+      put :update, params: { id: 1, harvest_job: { stop: true } }, format: 'js'
     end
   end
 end
