@@ -5,13 +5,15 @@
 # Supplejack was created by DigitalNZ at the National Library of NZ
 # and the Department of Internal Affairs. http://digitalnz.org/supplejack
 
-FactoryGirl.define do
+FactoryBot.define do
   factory :abstract_job do
-    start_time    Time.now
-    environment   "test"
+    start_time       Time.now
+    environment      "test"
 
     sequence(:parser_id)  {|n| "abc#{n}" }
     sequence(:version_id) {|n| "abc#{n}" }
     sequence(:user_id)    {|n| "abc#{n}" }
+
+    association :harvest_schedule, factory: :harvest_schedule
   end
 end
