@@ -7,7 +7,7 @@
 
 # Read about factories at https://github.com/thoughtbot/factory_girl
 
-FactoryGirl.define do
-  factory :preview do
+FactoryBot.define do
+  factory :collection_statistics, :class => 'CollectionStatistics' do
   end
 end
