@@ -5,7 +5,7 @@
 # Supplejack was created by DigitalNZ at the National Library of NZ
 # and the Department of Internal Affairs. http://digitalnz.org/supplejack
 
-FactoryGirl.define do
+FactoryBot.define do
   factory :harvest_job do
     limit         nil
     start_time    Time.now
@@ -14,5 +14,7 @@ FactoryGirl.define do
     sequence(:parser_id)  {|n| "abc#{n}" }
     sequence(:version_id) {|n| "abc#{n}" }
     sequence(:user_id)    {|n| "abc#{n}" }
+
+    association :harvest_schedule, factory: :harvest_schedule
   end
 end
