@@ -7,7 +7,7 @@
 
 # Read about factories at https://github.com/thoughtbot/factory_girl
 
-FactoryGirl.define do
+FactoryBot.define do
   factory :link_check_job do
     url "http://google.co.nz"
     record_id 123
