@@ -5,7 +5,7 @@
 # Supplejack was created by DigitalNZ at the National Library of NZ
 # and the Department of Internal Affairs. http://digitalnz.org/supplejack
 
-require 'spec_helper'
+require 'rails_helper'
 
 describe Parser do
 
@@ -31,8 +31,8 @@ describe Parser do
 
   describe "#last_harvested_at" do
     let!(:time) { Time.now }
-    let!(:job1) { FactoryGirl.create(:harvest_job, start_time: time - 1.day, parser_id: "12", status: "finished") }
-    let!(:job2) { FactoryGirl.create(:harvest_job, start_time: time - 2.day, parser_id: "12", status: "finished") }
+    let!(:job1) { FactoryBot.create(:harvest_job, start_time: time - 1.day, parser_id: "12", status: "finished") }
+    let!(:job2) { FactoryBot.create(:harvest_job, start_time: time - 2.day, parser_id: "12", status: "finished") }
 
     it "should return the last date a harvest job was run" do
       Timecop.freeze(time) do
