@@ -5,11 +5,11 @@
 # Supplejack was created by DigitalNZ at the National Library of NZ
 # and the Department of Internal Affairs. http://digitalnz.org/supplejack
 
-require 'spec_helper'
+require 'rails_helper'
 
 describe ApiDeleteWorker do
   let(:worker) { ApiDeleteWorker.new }
-  let(:job) { FactoryGirl.create(:harvest_job) }
+  let(:job) { FactoryBot.create(:harvest_job) }
 
   describe '#perform' do
     let(:success_response) do
