@@ -1,10 +1,9 @@
-# The Supplejack Worker code is Crown copyright (C) 2014, New Zealand Government, 
-# and is licensed under the GNU General Public License, version 3. 
-# See https://github.com/DigitalNZ/supplejack_worker for details. 
-# 
-# Supplejack was created by DigitalNZ at the National Library of NZ
-# and the Department of Internal Affairs. http://digitalnz.org/supplejack
+class ApplicationController < ActionController::API
+  def authenticate_user!
+    true
+  end
 
-class ApplicationController < ActionController::Base
-  protect_from_forgery
+  def current_user
+    User.first
+  end
 end
