@@ -8,7 +8,7 @@
 class LinkCheckRulesController < ApplicationController
 
   def index
-    @link_check_rules = params[:link_check_rule].present? ? LinkCheckRule.where(params[:link_check_rule]) : LinkCheckRule.all
+    @link_check_rules = params[:link_check_rule].present? ? LinkCheckRule.where(link_check_rule_params) : LinkCheckRule.all
     render json: @link_check_rules
   end
 
@@ -23,13 +23,19 @@ class LinkCheckRulesController < ApplicationController
   end
 
   def create
-    @link_check_rule = LinkCheckRule.create(params[:link_check_rule])
+    @link_check_rule = LinkCheckRule.create!(link_check_rule_params)
     render json: @link_check_rule
   end
 
   def update
     @link_check_rule = LinkCheckRule.find(params[:id])
-    @link_check_rule.update_attributes(params[:link_check_rule])
+    @link_check_rule.update_attributes(link_check_rule_params)
     render json: @link_check_rule
+  end
+
+  private
+
+  def link_check_rule_params
+    params.require(:link_check_rule).permit!
   end
 end
