@@ -1,12 +1,5 @@
-# The Supplejack Worker code is Crown copyright (C) 2014, New Zealand Government, 
-# and is licensed under the GNU General Public License, version 3. 
-# See https://github.com/DigitalNZ/supplejack_worker for details. 
-# 
-# Supplejack was created by DigitalNZ at the National Library of NZ
-# and the Department of Internal Affairs. http://digitalnz.org/supplejack
+# Load the Rails application.
+require_relative 'application'
 
-# Load the rails application
-require File.expand_path('../application', __FILE__)
-
-# Initialize the rails application
-HarvesterWorker::Application.initialize!
+# Initialize the Rails application.
+Rails.application.initialize!
