@@ -1,12 +1,6 @@
-# The Supplejack code is Crown copyright (C) 2014, New Zealand Government, 
-# and is licensed under the GNU General Public License, version 3. 
-# See https://github.com/DigitalNZ/supplejack_worker for details. 
-# 
-# Supplejack was created by DigitalNZ at the National Library of NZ
-# and the Department of Internal Affairs. http://digitalnz.org/supplejack
-
+# frozen_string_literal: true
 Kaminari.configure do |config|
-  config.default_per_page = 20
+  # config.default_per_page = 25
   # config.max_per_page = nil
   # config.window = 4
   # config.outer_window = 0
@@ -14,4 +8,5 @@ Kaminari.configure do |config|
   # config.right = 0
   # config.page_method_name = :page
   # config.param_name = :page
+  # config.params_on_first_page = false
 end
