@@ -5,51 +5,51 @@
 # Supplejack was created by DigitalNZ at the National Library of NZ
 # and the Department of Internal Affairs. http://digitalnz.org/supplejack
 
-require "spec_helper"
+require 'rails_helper'
 
 describe AbstractJob do
 
-  let(:job) { create(:abstract_job, parser_id: "12345", version_id: "666") }
+  let(:job) { create(:abstract_job, parser_id: '12345', version_id: '666') }
 
-  describe ".search" do
-    let!(:active_job) { create(:abstract_job, status: "active") }
+  describe '.search' do
+    let!(:active_job) { create(:abstract_job, status: 'active') }
 
-    it "returns all active harvest jobs" do
-      finished_job = create(:abstract_job, status: "finished")
-      AbstractJob.search("status" => "active").should eq [active_job]
+    it 'returns all active harvest jobs' do
+      create(:abstract_job, status: 'finished')
+      AbstractJob.search('status' => 'active').should eq [active_job]
     end
 
-    it "paginates through the records" do
+    it 'paginates through the records' do
       AbstractJob.should_receive(:page).with(2) { AbstractJob.unscoped }
-      AbstractJob.search("status" => "active", "page" => "2").to_a
+      AbstractJob.search('status' => 'active', 'page' => '2').to_a
     end
 
-    it "returns the recent harvest jobs first" do
-      active_job2 = create(:abstract_job, status: "active", start_time: Time.now + 5.seconds)
-      AbstractJob.search("status" => "active").first.should eq active_job2
+    it 'returns the recent harvest jobs first' do
+      active_job2 = create(:abstract_job, status: 'active', start_time: Time.now + 5.seconds)
+      AbstractJob.search('status' => 'active').first.should eq active_job2
     end
 
-    it "returns only test harvest jobs of a specific parser" do
-      job2 = create(:abstract_job, parser_id: "333", environment: "test")
-      AbstractJob.search("parser_id" => "333", "environment" => "test").should eq [job2]
+    it 'returns only test harvest jobs of a specific parser' do
+      job2 = create(:abstract_job, parser_id: '333', environment: 'test')
+      AbstractJob.search('parser_id' => '333', 'environment' => 'test').should eq [job2]
     end
 
-    it "limits the number of harvest jobs returned" do
-      job2 = create(:abstract_job, parser_id: "333", environment: "test", start_time: Time.now + 5.seconds)
-      AbstractJob.search("limit" => "1").to_a.size.should eq 1
+    it 'limits the number of harvest jobs returned' do
+      create(:abstract_job, parser_id: '333', environment: 'test', start_time: Time.now + 5.seconds)
+      AbstractJob.search('limit' => '1').to_a.size.should eq 1
     end
 
-    it "should find all harvest jobs either in staging or production" do
-      job1 = create(:abstract_job, parser_id: "333", environment: "staging", start_time: Time.now)
-      job2 = create(:abstract_job, parser_id: "334", environment: "production", start_time: Time.now + 2.seconds)
-      jobs = AbstractJob.search("environment" => ["staging", "production"]).to_a
+    it 'should find all harvest jobs either in staging or production' do
+      job1 = create(:abstract_job, parser_id: '333', environment: 'staging', start_time: Time.now)
+      job2 = create(:abstract_job, parser_id: '334', environment: 'production', start_time: Time.now + 2.seconds)
+      jobs = AbstractJob.search('environment' => ['staging', 'production']).to_a
       jobs.should include(job2)
       jobs.should include(job1)
     end
   end
 
-  describe ".clear_raw_data" do
-    it "should fetch harvest jobs older than a week" do
+  describe '.clear_raw_data' do
+    it 'should fetch harvest jobs older than a week' do
       AbstractJob.should_receive(:disposable) { [job] }
       job.should_receive(:clear_raw_data)
       AbstractJob.clear_raw_data
@@ -69,87 +69,87 @@ describe AbstractJob do
     end
   end
 
-  describe "#parser" do
+  describe '#parser' do
     let!(:version) { mock_model(ParserVersion).as_null_object }
 
-    it "returns @parser if @parser is set" do
+    it 'returns @parser if @parser is set' do
       job.instance_variable_set(:@parser, 'a parser!')
       job.parser.should eq 'a parser!'
     end
 
-    context "with version_id" do
-      it "finds the parser by id" do
-        ParserVersion.should_receive(:find).with("666", params: {parser_id: "12345"})
+    context 'with version_id' do
+      it 'finds the parser by id' do
+        ParserVersion.should_receive(:find).with('666', params: {parser_id: '12345'})
         job.parser
       end
     end
 
-    context "without version_id" do
+    context 'without version_id' do
       before(:each) do
-        job.version_id = ""
-        job.environment = "staging"
+        job.version_id = ''
+        job.environment = 'staging'
       end
 
-      it "finds the current parser version for the environment" do
-        ParserVersion.should_receive(:find).with(:one, from: :current, params: {parser_id: "12345", environment: "staging"}) { version }
+      it 'finds the current parser version for the environment' do
+        ParserVersion.should_receive(:find).with(:one, from: :current, params: {parser_id: '12345', environment: 'staging'}) { version }
         job.parser
       end
 
-      it "should set the version_id of the fetched version" do
-        ParserVersion.should_receive(:find) { mock_model(ParserVersion, id: "888").as_null_object }
+      it 'should set the version_id of the fetched version' do
+        ParserVersion.should_receive(:find) { mock_model(ParserVersion, id: '888').as_null_object }
         job.parser
-        job.version_id.should eq "888"
+        job.version_id.should eq '888'
       end
 
-      context "the environment is preview and parser_code is present" do
+      context 'the environment is preview and parser_code is present' do
         let(:parser) {double(:parser).as_null_object}
         before(:each) do
-          job.environment = "preview"
-          job.parser_code = "new code"
+          job.environment = 'preview'
+          job.parser_code = 'new code'
           Parser.stub(:find) {parser}
         end
 
-        it "finds the parser by id" do
-          Parser.should_receive(:find).with("12345")
+        it 'finds the parser by id' do
+          Parser.should_receive(:find).with('12345')
           job.parser
         end
 
-        it "sets the parser content to parser_code" do
-          parser.should_receive(:content=).with("new code")
+        it 'sets the parser content to parser_code' do
+          parser.should_receive(:content=).with('new code')
           job.parser
         end
       end
     end
 
-    context "without version_id and environment" do
+    context 'without version_id and environment' do
       before do
-        job.version_id = ""
+        job.version_id = ''
         job.environment = nil
       end
 
-      it "finds the parser by id" do
-        Parser.should_receive(:find).with("12345")
+      it 'finds the parser by id' do
+        Parser.should_receive(:find).with('12345')
         job.parser
       end
     end
   end
 
-  describe "#required_enrichments" do
-    it "should return an array of enrichments with required: true" do
+  describe '#required_enrichments' do
+    it 'should return an array of enrichments with required: true' do
       job.stub_chain(:parser, :enrichment_definitions).and_return({ndha_rights: {required_for_active_record: true}, thumbnails: {} })
       job.required_enrichments.should eq [:ndha_rights]
     end
   end
 
-  describe "status" do
-    describe "initial" do
-      it "should set the initial status to active" do
-        job.ready?.should be_true
+  describe 'status' do
+    describe 'initial' do
+      it 'should set the initial status to active' do
+        job.ready?.should be_truthy
       end
     end
 
-    describe "start!" do
-      it "should set the start time to now" do
+    describe 'start!' do
+      it 'should set the start time to now' do
         time = Time.now
         Timecop.freeze(time) do
           job.start!
@@ -157,34 +157,34 @@ describe AbstractJob do
         end
       end
 
-      it "should set the status to active" do
+      it 'should set the status to active' do
         job.start!
-        job.active?.should be_true
+        job.active?.should be_truthy
       end
 
-      it "should set the records count to 0" do
+      it 'should set the records count to 0' do
         job.start!
         job.records_count.should eq 0
       end
 
-      it "should set the processed count to 0" do
+      it 'should set the processed count to 0' do
         job.start!
         job.processed_count.should eq 0
       end
 
-      it "should save the job" do
+      it 'should save the job' do
         job.should_receive(:save)
         job.start!
       end
     end
 
-    describe "finish!" do
-      it "should set the status to finished" do
+    describe 'finish!' do
+      it 'should set the status to finished' do
         job.finish!
-        job.finished?.should be_true
+        job.finished?.should be_truthy
       end
 
-      it "should set the end time to now" do
+      it 'should set the end time to now' do
         time = Time.now
         Timecop.freeze(time) do
           job.finish!
@@ -192,30 +192,30 @@ describe AbstractJob do
         end
       end
 
-      it "should set the throughput" do
+      it 'should set the throughput' do
         job.should_receive(:calculate_throughput)
         job.finish!
       end
 
-      it "should set the errors count" do
+      it 'should set the errors count' do
         job.should_receive(:calculate_errors_count)
         job.finish!
       end
 
-      it "should save the job" do
+      it 'should save the job' do
         job.should_receive(:save)
         job.finish!
       end
     end
 
-    describe "error!" do
-      it "should set the status to failed" do
+    describe 'error!' do
+      it 'should set the status to failed' do
         job.error!
-        job.failed?.should be_true
+        job.failed?.should be_truthy
       end
 
-      context "start time" do
-        it "should set the start time to now if not set" do
+      context 'start time' do
+        it 'should set the start time to now if not set' do
           time = Time.now
           Timecop.freeze(time) do
             job.error!
@@ -223,7 +223,7 @@ describe AbstractJob do
           end
         end
 
-        it "should not set the start time to now if set" do
+        it 'should not set the start time to now if set' do
           Timecop.freeze(Time.now - 1.hour) do
             job.start!
           end
@@ -236,7 +236,7 @@ describe AbstractJob do
         end
       end
 
-      it "should set the end time to now" do
+      it 'should set the end time to now' do
         job.start!
         time = Time.now
         Timecop.freeze(time) do
@@ -245,28 +245,28 @@ describe AbstractJob do
         end
       end
 
-      it "should set the errors count" do
+      it 'should set the errors count' do
         job.should_receive(:calculate_errors_count)
         job.error!
       end
 
-      it "should save the job" do
+      it 'should save the job' do
         job.should_receive(:save)
         job.error!
       end
     end
 
-    describe "stop!" do
+    describe 'stop!' do
       before(:each) do
         job.start!
       end
 
-      it "should set the status to failed" do
+      it 'should set the status to failed' do
         job.stop!
-        job.stopped?.should be_true
+        job.stopped?.should be_truthy
       end
 
-      it "should set the end time to now" do
+      it 'should set the end time to now' do
         time = Time.now
         Timecop.freeze(time) do
           job.stop!
@@ -274,58 +274,58 @@ describe AbstractJob do
         end
       end
 
-      it "should save the job" do
+      it 'should save the job' do
         job.should_receive(:save)
         job.stop!
       end
     end
   end
 
-  describe "test?" do
-    it "returns true" do
-      job.environment = "test"
-      job.test?.should be_true
+  describe 'test?' do
+    it 'returns true' do
+      job.environment = 'test'
+      job.test?.should be_truthy
     end
 
-    it "returns false" do
-      job.environment = "staging"
-      job.test?.should be_false
-    end
-  end
-
-  describe "#preview?" do
-    it "returns true" do
-      job.environment = "preview"
-      job.preview?.should be_true
-    end
-
-    it "returns false" do
-      job.environment = "staging"
-      job.preview?.should be_false
+    it 'returns false' do
+      job.environment = 'staging'
+      job.test?.should be_falsey
     end
   end
 
-  describe "calculate_throughput" do
+  describe '#preview?' do
+    it 'returns true' do
+      job.environment = 'preview'
+      job.preview?.should be_truthy
+    end
+
+    it 'returns false' do
+      job.environment = 'staging'
+      job.preview?.should be_falsey
+    end
+  end
+
+  describe 'calculate_throughput' do
     before(:each) do
       job.end_time = Time.now
-      job.status = "finished"
+      job.status = 'finished'
     end
 
-    it "should calculate the average record time" do
+    it 'should calculate the average record time' do
       job.records_count = 100
       job.stub(:duration) { 100 }
       job.calculate_throughput
       job.throughput.should eq 1.0
     end
 
-    it "returns 0 when records harvested is 0" do
+    it 'returns 0 when records harvested is 0' do
       job.records_count = 0
       job.stub(:duration) { 100 }
       job.calculate_throughput
       job.throughput.should eq 0
     end
 
-    it "should not return NaN" do
+    it 'should not return NaN' do
       job.records_count = 0
       job.stub(:duration) { 0.0 }
       job.calculate_throughput
@@ -333,22 +333,22 @@ describe AbstractJob do
     end
   end
 
-  describe ".jobs_since" do
+  describe '.jobs_since' do
 
-    let!(:finished_job) { create(:abstract_job, status: "finished", start_time: (DateTime.now - 1), environment: "staging" ) }
+    let!(:finished_job) { create(:abstract_job, status: 'finished', start_time: (DateTime.now - 1), environment: 'staging' ) }
 
-    it "returns a count of harvest jobs in the last 2 days" do
-      old_finished_job = create(:abstract_job, status: "finished", start_time: (DateTime.now - 3), environment: "staging" )
+    it 'returns a count of harvest jobs in the last 2 days' do
+      old_finished_job = create(:abstract_job, status: 'finished', start_time: (DateTime.now - 3), environment: 'staging' )
       since = DateTime.now - 2
       AbstractJob.jobs_since( {'datetime'=>since.to_s, 'environment'=>'staging', 'status'=>'finished'} ).should eq [finished_job]
     end
 
   end
 
-  describe "#duration" do
+  describe '#duration' do
     let!(:time) { Time.now }
 
-    it "should return the duration in seconds" do
+    it 'should return the duration in seconds' do
       job.start_time = time - 10.seconds
       job.end_time = time
       job.save
@@ -356,17 +356,17 @@ describe AbstractJob do
       job.duration.should eq 10.0
     end
 
-    it "returns nil start_time is nil" do
+    it 'returns nil start_time is nil' do
       job.start_time = nil
       job.duration.should be_nil
     end
 
-    it "returns nil end_time is nil" do
+    it 'returns nil end_time is nil' do
       job.end_time = nil
       job.duration.should be_nil
     end
 
-    it "returns the proper duration" do
+    it 'returns the proper duration' do
       time = Time.now
       Timecop.freeze(time) do
         job = create(:abstract_job, start_time: time)
@@ -376,58 +376,58 @@ describe AbstractJob do
     end
   end
 
-  describe "total_errors_count" do
-    it "should return a sum of failed and invalid records" do
+  describe 'total_errors_count' do
+    it 'should return a sum of failed and invalid records' do
       job.stub(:invalid_records) { double(:array, count: 10) }
       job.stub(:failed_records) { double(:array, count: 20) }
       job.total_errors_count.should eq 30
     end
   end
 
-  describe "#errors_over_limit?" do
-    context "errors count over 100" do
+  describe '#errors_over_limit?' do
+    context 'errors count over 100' do
       before { job.stub(:total_errors_count) { 101 } }
 
-      it "should return true" do
-        job.errors_over_limit?.should be_true
+      it 'should return true' do
+        job.errors_over_limit?.should be_truthy
       end
     end
 
-    context "errors count under 100" do
+    context 'errors count under 100' do
       before { job.stub(:total_errors_count) { 99 } }
 
-      it "should return false" do
-        job.errors_over_limit?.should be_false
+      it 'should return false' do
+        job.errors_over_limit?.should be_falsey
       end
     end
   end
 
-  describe "clear_raw_data" do
-    it "should remove invalid records" do
-      job.invalid_records.create(raw_data: "Wrong", errors_messages: [])
+  describe 'clear_raw_data' do
+    it 'should remove invalid records' do
+      job.invalid_records.create(raw_data: 'Wrong', errors_messages: [])
       job.clear_raw_data
       job.reload
       job.invalid_records.count.should eq 0
     end
 
-    it "should remove failed records" do
-      job.failed_records.create(message: "Hi")
+    it 'should remove failed records' do
+      job.failed_records.create(message: 'Hi')
       job.clear_raw_data
       job.reload
       job.failed_records.count.should eq 0
     end
   end
 
-  describe "#increment_records_count!" do
-    it "should increment the records count" do
+  describe '#increment_records_count!' do
+    it 'should increment the records count' do
       job.increment_records_count!
       job.reload
       job.records_count.should eq 1
     end
   end
 
-  describe "#increment_processed_count!" do
-    it "should increment the records count" do
+  describe '#increment_processed_count!' do
+    it 'should increment the records count' do
       job.increment_processed_count!
       job.reload
       job.processed_count.should eq 1
